move the Witness builder into a separate module
